requirements.txt
streamlit>=1.54.0
pdfplumber>=0.11.4
pymupdf>=1.24.13
numpy>=1.26.0
sentence-transformers>=3.2.0
chromadb>=1.5.0
pydantic>=2.12.0
google-genai>=1.0.0
python-dotenv>=1.0.1
pytest>=8.3.0
